pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "quest-wrist-tracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiortc>=1.14.0",
    "av>=16.1.0",
    "hand-tracking-sdk>=1.1.0",
    "imageio>=2.37.0",
    "imageio-ffmpeg>=0.6.0",
    "numpy>=2.4.2",
    "matplotlib>=3.10.8",
    "pillow>=11.2.1",
    "pyarrow>=21.0.0",
    "websockets>=16.0",
]

[tool.setuptools]
py-modules = []
